% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioElement.R
\name{bindModuleWeight}
\alias{bindModuleWeight}
\title{A bindModuleWeight Function
bind}
\usage{
bindModuleWeight(weight_list, parameter_set)
}
\arguments{
\item{weight_list}{is list of asset parameters}

\item{parameter_set}{is xts object}
}
\description{
A bindModuleWeight Function
bind
}
\examples{
#' bindModuleWeight(weight_list, parameter_set)
}
\keyword{say}
