% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{changeStr}
\alias{changeStr}
\title{A changeStr Function
changeStr}
\usage{
changeStr(string)
}
\arguments{
\item{string}{strings}
}
\description{
A changeStr Function
changeStr
}
\examples{
#' changeStr(string)
}
\keyword{changeStr}
