% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{checkDateFormat}
\alias{checkDateFormat}
\title{A checkDateFormat Function
checkDateFormat}
\usage{
checkDateFormat(date_index)
}
\arguments{
\item{date_index}{is xts object}
}
\description{
A checkDateFormat Function
checkDateFormat
}
\examples{
#' checkDateFormat(date_index)
}
\keyword{checkDateFormat}
