% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioAnal.R
\name{cumContribution}
\alias{cumContribution}
\title{A cumContribution Function
calculate cumulative Contribution}
\usage{
cumContribution(wealthIndex, contribution)
}
\arguments{
\item{wealthIndex}{is xts object from return.portfolio}

\item{contribution}{is xts object from return.portfolio}
}
\description{
A cumContribution Function
calculate cumulative Contribution
}
\examples{
#' cumContribution(wealthIndex, contribution)
}
\keyword{cumContribution}
