% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateIndexing.R
\name{dayPick}
\alias{dayPick}
\title{A dayPick Function
calculate dayPick}
\usage{
dayPick(
  date_index,
  split_grid = c("week", "month"),
  numb = 21,
  days_in = 21,
  n = 1
)
}
\arguments{
\item{date_index}{is xts object}

\item{split_grid}{is number of days to average}

\item{numb}{is number of days to average}

\item{days_in}{is number of days to average}

\item{n}{is number of days to average}
}
\description{
A dayPick Function
calculate dayPick
}
\examples{
#' dayPickInMonth(date_index, numb = 21, days_in_month = 21, n_month=1)
}
\keyword{dayPickInMonth}
