% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateIndexing.R
\name{dayPickInMonth}
\alias{dayPickInMonth}
\title{A dayPickInMonth Function
calculate dayPickInMonth}
\usage{
dayPickInMonth(date_index, numb = 21, days_in_month = 21, n_month = 1)
}
\arguments{
\item{date_index}{is xts object}

\item{numb}{is number of days to average}

\item{days_in_month}{is number of days to average}

\item{n_month}{is number of days to average}
}
\description{
A dayPickInMonth Function
calculate dayPickInMonth
}
\examples{
#' dayPickInMonth(date_index, numb = 21, days_in_month = 21, n_month=1)
}
\keyword{dayPickInMonth}
