% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateIndexing.R
\name{dayPickInWeek}
\alias{dayPickInWeek}
\title{A dayPickInWeek Function
calculate dayPickInWeek}
\usage{
dayPickInWeek(date_index, numb = 5, days_in_week = 5, n_week = 1)
}
\arguments{
\item{date_index}{is xts object}

\item{numb}{is number of days to average}

\item{days_in_week}{is number of days to average}

\item{n_week}{is number of days to average}
}
\description{
A dayPickInWeek Function
calculate dayPickInWeek
}
\examples{
#' dayPickInWeek(date_index, numb = 5, days_in_week = 5, n_week = 1)
}
\keyword{dayPickInWeek}
