% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioAnal.R
\name{Drawdowns}
\alias{Drawdowns}
\title{A Drawdowns Function
calculate cumulative Contribution}
\usage{
Drawdowns(daily_return)
}
\arguments{
\item{daily_return}{is xts object from return.portfolio}
}
\description{
A Drawdowns Function
calculate cumulative Contribution
}
\examples{
#' Drawdowns(daily_return)
}
\keyword{Drawdowns}
