% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacticalIndicator.R
\name{endMonthRollingReturn}
\alias{endMonthRollingReturn}
\title{A endMonthRollingReturn Function
calculate endMonthRollingReturn}
\usage{
endMonthRollingReturn(data, roll_month = 24)
}
\arguments{
\item{data}{is xts}

\item{roll_month}{default is 24}
}
\description{
A endMonthRollingReturn Function
calculate endMonthRollingReturn
}
\examples{
#' endMonthRollingReturn(data, roll_month = 24)
}
\keyword{endMonthRollingReturn}
