% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagMeasure.R
\name{laggedCorrelation}
\alias{laggedCorrelation}
\title{A laggedCorrelation Function
calculate laggedCorrelation}
\usage{
laggedCorrelation(price_table)
}
\arguments{
\item{price_table}{is xts object}
}
\description{
A laggedCorrelation Function
calculate laggedCorrelation
}
\examples{
#' laggedCorrelation(price_table)
}
\keyword{laggedCorrelation}
