% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagMeasure.R
\name{laggedPrice}
\alias{laggedPrice}
\title{A laggedPrice Function
calculate laggedPrice}
\usage{
laggedPrice(price_1, price_2, merge_by = 1)
}
\arguments{
\item{price_1}{is xts object}

\item{price_2}{is xts object}

\item{merge_by}{is xts object}
}
\description{
A laggedPrice Function
calculate laggedPrice
}
\examples{
#' laggedPrice(price_1, price_2, merge_by = 1)
}
\keyword{laggedPrice}
