% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacticalIndicator.R
\name{momentumRank}
\alias{momentumRank}
\title{A momentumRank Function
calculate momentumRank}
\usage{
momentumRank(momentum_score, rank_method = "min")
}
\arguments{
\item{momentum_score}{is score}

\item{rank_method}{default is 'min'}
}
\description{
A momentumRank Function
calculate momentumRank
}
\examples{
#' momentumRank(momentum_score=NULL, rank_method = "min")
}
\keyword{momentumRank}
