% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacticalIndicator.R
\name{movingAverage}
\alias{movingAverage}
\title{A movingAverage Function
calculate moving average}
\usage{
movingAverage(price, n)
}
\arguments{
\item{price}{is xts object}

\item{n}{is number of days to average}
}
\description{
A movingAverage Function
calculate moving average
}
\examples{
#' movingAverage(price, n = 200)
}
\keyword{say}
