% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioAnal.R
\name{pfStatistics}
\alias{pfStatistics}
\title{A pfStatistics Function}
\usage{
pfStatistics(wealth_index, cal_obj = "all")
}
\arguments{
\item{wealth_index}{is xts object}

\item{cal_obj}{default is 'all'}
}
\description{
A pfStatistics Function
}
\examples{
#' pfStatistics(wealth_index, cal_obj = "all")
}
\keyword{pfStatistics}
