% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioAnal.R
\name{pfStatsTable}
\alias{pfStatsTable}
\title{A pfStatsTable Function}
\usage{
pfStatsTable(wealth_index_table, cal_obj = "all")
}
\arguments{
\item{wealth_index_table}{is xts object}

\item{cal_obj}{default is 'all'}
}
\description{
A pfStatsTable Function
}
\examples{
#' pfStatsTable(wealth_index_table, cal_obj = "all")
}
\keyword{pfStatistics}
