% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacticalIndicator.R
\name{priceRankCorr}
\alias{priceRankCorr}
\title{A priceRankCorr Function
calculate priceRankCorr}
\usage{
priceRankCorr(price, mom_period = c(1:12), split_grid = "month", ...)
}
\arguments{
\item{price}{is xts object}

\item{mom_period}{vector of month}

\item{split_grid}{date grid, default is 'month'}

\item{...}{parameters for dayPick}
}
\description{
A priceRankCorr Function
calculate priceRankCorr
}
\examples{
#' priceRankCorr(price, mom_period= c(1:12), split_grid = "month")
}
\keyword{priceRankCorr}
