% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{readCSVtoXTS}
\alias{readCSVtoXTS}
\title{A readCSVtoXTS Function
scalePrice}
\usage{
readCSVtoXTS(file_path, col_select = c("date", "close"))
}
\arguments{
\item{file_path}{is xts object}

\item{col_select}{is xts object}
}
\description{
A readCSVtoXTS Function
scalePrice
}
\examples{
#' readCSVtoXTS(file_path, col_select = c("date", "close"))
}
\keyword{scalePrice}
