% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioAnal.R
\name{restrictedPartition}
\alias{restrictedPartition}
\title{A restrictedPartition Function
restrictedPartition}
\usage{
restrictedPartition(total, minDeposits, depositUnits, allowZero = FALSE)
}
\arguments{
\item{total}{is scalar}
}
\description{
A restrictedPartition Function
restrictedPartition
}
\examples{
#' restrictedPartition(total= 100, minDeposits=c(0,0,0), depositUnits= c(5,5,5))
}
\keyword{weight}
