% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioAnal.R
\name{returnPortfolio}
\alias{returnPortfolio}
\title{A returnPortfolio Function}
\usage{
returnPortfolio(
  R,
  weights = NULL,
  wealth.index = TRUE,
  contribution = TRUE,
  geometric = TRUE,
  rebalance_on = c(NA, "years", "quarters", "months", "weeks", "days"),
  value = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{R}{is reuturns of assets}

\item{weights}{weight xts}

\item{wealth.index}{default is TRUE}

\item{contribution}{default is TRUE}

\item{geometric}{default is TRUE}

\item{rebalance_on}{default NA}

\item{value}{default is 1, starting value}

\item{verbose}{default is TRUE}

\item{...}{Return.portfolio.geometric}
}
\description{
A returnPortfolio Function
}
\keyword{returnPortfolio}
