% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{save_list_to_xlsx}
\alias{save_list_to_xlsx}
\title{A savelisttoxlsx Function
scalePrice}
\usage{
save_list_to_xlsx(data_list, file_name)
}
\arguments{
\item{data_list}{list of data.frames}

\item{file_name}{save file directory}
}
\description{
A savelisttoxlsx Function
scalePrice
}
\examples{
#' save_list_to_xlsx(data_list, file_name)
}
\keyword{list}
