% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{A say Function}
\usage{
say(state = TRUE)
}
\arguments{
\item{state}{is true default}
}
\description{
This function allows you to express your saying bye
}
\examples{
say()
}
\keyword{say}
