% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{sayROC}
\alias{sayROC}
\title{A sayROC Function}
\usage{
sayROC(state = TRUE)
}
\arguments{
\item{state}{is true default}
}
\description{
This function allows you to express your saying bye
}
\examples{
sayROC()
}
\keyword{say}
