% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{scalePrice}
\alias{scalePrice}
\title{A scalePrice Function
scalePrice}
\usage{
scalePrice(price)
}
\arguments{
\item{price}{is xts object}
}
\description{
A scalePrice Function
scalePrice
}
\examples{
#' checkDateFormat(date_index)
}
\keyword{scalePrice}
