% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioAnal.R
\name{skipNA}
\alias{skipNA}
\title{A skipNA Function
skipNA}
\usage{
skipNA(data, FUN = NULL)
}
\arguments{
\item{data}{is xts object from return.portfolio}

\item{FUN}{is xts object from return.portfolio}
}
\description{
A skipNA Function
skipNA
}
\examples{
#' skipNA(data, FUN = NULL)
}
\keyword{Drawdowns}
