% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacticalIndicator.R
\name{tsMomentumScore}
\alias{tsMomentumScore}
\title{A tsMomentumScore Function
calculate tmomentum score}
\usage{
tsMomentumScore(price, mom_period, split_grid = "month", ...)
}
\arguments{
\item{price}{is xts object}

\item{mom_period}{vector of month}

\item{split_grid}{date grid, default is 'month'}

\item{...}{parameters for dayPick}
}
\description{
A tsMomentumScore Function
calculate tmomentum score
}
\examples{
#' tsMomentumScore(price, mom_period= c(1:12), split_grid = "month")
}
\keyword{tsMomentumScore}
